refactor: llm module addition
