+++ reasoners/sequential/interface.py
@@ -5,7 +5,7 @@ from typing import Any, Deque, Dict
 
 from memory.base_memory import BaseMemory
 from tools.interface import ToolInterface
-from utils.llm import BaseLLM
+from llm.base_llm import BaseLLM
 from reasoners.models import ReasonerState, Step
